+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/ai_assistant/helpers.ts
