+++ x-pack/test/api_integration/deployment_agnostic/configs/serverless/oblt.index.ts
@@ -19,5 +19,6 @@ export default function ({ loadTestFile }: DeploymentAgnosticFtrProviderContext)
     loadTestFile(require.resolve('../../apis/saved_objects_management'));
     loadTestFile(require.resolve('../../apis/observability/slo'));
     loadTestFile(require.resolve('../../apis/observability/synthetics'));
+    loadTestFile(require.resolve('../../apis/observability/ai_assistant'));
   });
 }
