+++ x-pack/test/api_integration/deployment_agnostic/services/deployment_agnostic_services.ts
@@ -21,7 +21,7 @@ export const deploymentAgnosticServices = _.pick(apiIntegrationServices, [
   'indexPatterns',
   'ingestPipelines',
   'kibanaServer',
-  // 'ml', depends on 'esDeleteAllIndices', can we make it deployment agnostic?
+  'ml',
   'randomness',
   'retry',
   'security',
